+++ src/components/item/AddButton.tsx
@@ -1,33 +1,14 @@
 import React from "react";
 import { StyleSheet, Text, View } from "react-native";
 import { TouchableOpacity } from "react-native-gesture-handler";
-import { useDispatch } from "react-redux";
-import { addContent } from "../../modules/store/store";
-import useGetDoList from "../../hooks/useGetDoList";
 
 export default function AddButton({
   newContent,
-  setAddText,
+  onAddContent,
 }: {
   newContent: string;
-  setAddText: any;
+  onAddContent: any;
 }) {
-  const dispatch = useDispatch();
-  const { toDoList } = useGetDoList();
-
-  const onAddContent = (content: string) => {
-    if (!content) return;
-    const newDoList = [
-      ...toDoList,
-      {
-        text: content,
-        checked: false,
-      },
-    ];
-    dispatch(addContent(newDoList));
-    setAddText("");
-  };
-
   return (
     <TouchableOpacity onPress={() => onAddContent(newContent)}>
       <View style={[styles.container]}>
