+++ src/views/screen/DoneList.tsx
@@ -1,10 +1,6 @@
-import React, { useState, useEffect } from "react";
-import { StyleSheet, ScrollView } from "react-native";
-import HeaderBar from "../../components/bar/HeaderBar";
-import { backColor } from "../../styles/common";
-import { useGetActiveTab } from "../../hooks/useGetDoList";
-import ToDoListContainer from "../../components/organizm/ToDoListContainer";
-import BottomAddBar from "../../components/bar/BottomAddBar";
+import React from "react";
+import { ScrollView } from "react-native";
+import ToDoListContainer from "../../components/organizm/TodoListContainer";
 
 export default function DoneList() {
   return (
@@ -13,10 +9,3 @@ export default function DoneList() {
     </ScrollView>
   );
 }
-
-const styles = StyleSheet.create({
-  topContainer: {
-    flex: 1,
-    backgroundColor: backColor.default,
-  },
-});
